pyproject.toml
[project]
name = "checar-video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "google-generativeai>=0.8.4",
    "langchain-groq>=0.2.4",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "whisper>=1.1.10",
]
